change float to type double

+++ include/ast/primary/ast_float.hpp
@@ -12,14 +12,14 @@ namespace ast {
 class Float : public Node 
 {
     public:
-        Float(float value);
+        Float(double value);
 
         virtual void EmitRISCV(std::ostream &stream, const std::string &dst_reg, Context &context) const override;
         virtual void Print(std::ostream &stream, int indent) const override;
 
 
     private:
-        float value_;
+        double value_;
 };
 
 } // namespace ast
